Change float to int in query_click, since click events use integer coordinates. Change draw to be a method of Button class that takes window as argument. Organize buttons and textfields in vectors. Change textfields to only accept digits.

+++ clickable.h
@@ -3,5 +3,5 @@
 class Clickable {
 public:
     virtual ~Clickable() = default;
-    virtual void query_click(float x, float y) = 0;
+    virtual void query_click(int x, int y) = 0;
 };
